Modules folder renamed to mod (#31)
