+++ cmd/honeypot/main.go
@@ -4,8 +4,8 @@ import (
 	"github.com/gliderlabs/ssh"
 	"github.com/kyberorg/honeypot/cmd/honeypot/config"
 	"github.com/kyberorg/honeypot/cmd/honeypot/dto"
-	"github.com/kyberorg/honeypot/cmd/honeypot/modules/prom"
-	"github.com/kyberorg/honeypot/cmd/honeypot/modules/rawmetrics"
+	"github.com/kyberorg/honeypot/cmd/honeypot/mod/prom"
+	"github.com/kyberorg/honeypot/cmd/honeypot/mod/rawmetrics"
 	"github.com/kyberorg/honeypot/cmd/honeypot/sshutil"
 	"github.com/kyberorg/honeypot/cmd/honeypot/util"
 	"github.com/kyberorg/honeypot/cmd/honeypot/writer"
